requirements.txt
backports-abc==0.4
decorator==4.0.6
praw==3.5.0
PyYAML==3.11
requests==2.8.1
six==1.10.0
tornado==4.3
update-checker==0.11
wheel==0.24.0
arrow
